Introduce BaseRESTTable as the single SupportsReadRestrictions carrier
The previous shape had a semantic lie in the type system: BaseTable carried
a ReadRestrictions field (to support the LOCAL scan-planning mode fallback
in RESTSessionCatalog) but deliberately did not implement
SupportsReadRestrictions. Only RESTTable — constructed in SERVER
scan-planning mode — advertised the marker. The default LOCAL mode
returned BaseTable, so 'instanceof SupportsReadRestrictions' quietly
returned false and restrictions were dropped.

Fix by introducing BaseRESTTable as an intermediate class in core/rest/:

  BaseTable      (core, generic; no restriction concept at all)
    └─ BaseRESTTable      (core/rest; field + implements marker)
         └─ RESTTable     (core/rest; adds scan planning)

BaseTable goes back to its pre-restrictions shape — removing the 4-arg
constructor, the field, and the readRestrictions() method. Non-REST
catalogs (Hadoop, Hive, Glue, JDBC, Nessie) that construct BaseTable
no longer advertise a capability they have no pathway to produce.

BaseRESTTable carries the field and implements SupportsReadRestrictions
directly. RESTSessionCatalog's three fallback sites (line 598 for LOCAL
loadTable, line 747 for commit, line 1020 for another commit path) now
construct BaseRESTTable instead of BaseTable, so restrictions flow
correctly in all modes.

RESTTable extends BaseRESTTable and inherits both the field and marker —
drops the redundant 'implements SupportsReadRestrictions' declaration.

Consumer code is unchanged:
  if (table instanceof SupportsReadRestrictions r) {
    r.readRestrictions().ifPresent(...);
  }

Now returns true only for REST-loaded tables, never for Hive/Hadoop/etc.
