+++ core/src/main/java/org/apache/iceberg/rest/RESTTable.java
@@ -21,20 +21,17 @@ package org.apache.iceberg.rest;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Supplier;
-import org.apache.iceberg.BaseTable;
 import org.apache.iceberg.BatchScan;
 import org.apache.iceberg.BatchScanAdapter;
 import org.apache.iceberg.ImmutableTableScanContext;
 import org.apache.iceberg.SupportsDistributedScanPlanning;
-import org.apache.iceberg.SupportsReadRestrictions;
 import org.apache.iceberg.TableOperations;
 import org.apache.iceberg.TableScan;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.metrics.MetricsReporter;
 import org.apache.iceberg.rest.restrictions.ReadRestrictions;
 
-class RESTTable extends BaseTable
-    implements SupportsDistributedScanPlanning, SupportsReadRestrictions {
+class RESTTable extends BaseRESTTable implements SupportsDistributedScanPlanning {
   private final RESTClient client;
   private final Supplier<Map<String, String>> headers;
   private final MetricsReporter reporter;
